update admin system

+++ app.py
@@ -48,7 +48,8 @@ def index():
         leaders = getscores()
         #print(sql,"sql")
         cliques=getscores2()
-        return render_template("index.html",score=sql,leaders = leaders, cliques=cliques )
+        adminstatus = isadmin()
+        return render_template("index.html",score=sql,leaders = leaders, cliques=cliques, adminstatus=adminstatus)
     else:
         total = db.session.execute(text("select sum(score) from scores")).fetchone()[0]
         if not total:
@@ -56,7 +57,8 @@ def index():
         sql = getthing()
         leaders = getscores()
         cliques=getscores2()
-        return render_template("index.html",score = sql,leaders = leaders,cliques=cliques,total=total)
+        adminstatus = isadmin()
+        return render_template("index.html",score = sql,leaders = leaders,cliques=cliques,total=total, adminstatus=adminstatus)
 
 def cliquescorechanger(userid,score):
     try:
@@ -75,9 +77,21 @@ def getthing(): #gets user score
         return 9999
 
 def isadmin(): #checks admin status
-    if db.session.execute(text("select id from users where username=:username"),{"username":session["username"]}).fetchone() in db.session.execute(text("select user_id from admins")).fetchall():
-        return True
-    return False
+    try:
+        if db.session.execute(text("select id from users where username=:username"),{"username":session["username"]}).fetchone() in db.session.execute(text("select user_id from admins")).fetchall():
+            return True
+        return False
+    except:
+        return False
+    
+def isbigboss():
+    try:
+        userid = db.session.execute(text("select id from users where username=:username"),{"username":session["username"]}).fetchone()[0]
+        if db.session.execute(text("select bigboss from admins where user_id=:userid"),{"userid":userid}).fetchone()[0] == True:
+            return True
+        return False
+    except:
+        return False
 
 def getuser(): #gets user id
     try:
@@ -126,14 +140,12 @@ def cliquescorehelper(clique):
     else:
         return False
     
-@app.route("/admin",methods=["POST","GET"])
+@app.route("/admin",methods=["GET"])
 def admin():
-    if request.method == "POST":
-        redirect("/")
-    else:
-        if isadmin() == False:
-            return redirect("/")
-        return render_template("admin.html")
+    if isadmin() == False:
+        return redirect("/")
+    bigboss = isbigboss()
+    return render_template("admin.html",bigboss=bigboss)
     
 @app.route("/deluser", methods=["POST"])
 def deluser():
@@ -144,6 +156,9 @@ def deluser():
             flash("whoopsie >__< (user not found)","warning")
             return redirect("/admin")
         userid = userid[0]
+        if (userid,) in db.session.execute(text("select user_id from admins")).fetchall():
+            flash("whoopsie >__< (can't delete admin)","warning")
+            return redirect("/admin")
         db.session.execute(text("delete from users where id=:id"),{"id":userid})
         db.session.execute(text("delete from scores where user_id=:user_id"),{"user_id":userid})
         db.session.execute(text("delete from cliques where user_id=:user_id"),{"user_id":userid})
@@ -173,6 +188,43 @@ def delclique():
 
     flash("aborted >__< (check the box, yeah?)","warning")
     return redirect("/admin")
+
+
+@app.route("/deladmin", methods=["POST"])
+def deladmin():
+    for i in request.form.getlist("yeahimsure"):
+        user = request.form["user"]
+        userid = db.session.execute(text("select id from users where username=:user"),{"user":user}).fetchone()
+        if userid == None:
+            flash("whoopsie >__< (user not found)","warning")
+            return redirect("/admin")
+        userid = userid[0]
+        db.session.execute(text("delete from admins where user_id=:user_id"),{"user_id":userid})
+        db.session.commit()
+        flash("yay ^v^ (admin privliges revoked)","warning")
+        return redirect("/admin")
+
+    flash("aborted >__< (check the box, yeah?)","warning")
+    return redirect("/admin")
+
+@app.route("/giveadmin", methods=["POST"])
+def giveadmin():
+    for i in request.form.getlist("yeahimsure"):
+        user = request.form["user"]
+        userid = db.session.execute(text("select id from users where username=:user"),{"user":user}).fetchone()
+        if userid == None:
+            flash("whoopsie >__< (user not found)","warning")
+            return redirect("/admin")
+        userid = userid[0]
+        db.session.execute(text("insert into admins (user_id,bigboss) values (:user_id,False) "),{"user_id":userid})
+        db.session.commit()
+        flash(f"yay ^v^ (admin privliges given to {user})","warning")
+        return redirect("/admin")
+
+    flash("aborted >__< (check the box, yeah?)","warning")
+    return redirect("/admin")
+
+
 
 @app.route("/login",methods=["POST","GET"])
 def login():
